+++ apps/cowswap-frontend/src/modules/appData/utils/buildAppData.ts
@@ -1,6 +1,5 @@
-import { stringifyDeterministic } from '@cowprotocol/app-data'
 import { UtmParams } from '@cowprotocol/common-utils'
-import { SupportedChainId } from '@cowprotocol/cow-sdk'
+import { stringifyDeterministic, SupportedChainId } from '@cowprotocol/cow-sdk'
 
 import { metadataApiSDK } from 'cowSdk'
 import { keccak256, toUtf8Bytes } from 'ethers/lib/utils'