+++ apps/cowswap-frontend/src/modules/notifications/containers/ConnectTelegram.tsx
@@ -1,12 +1,12 @@
-import { useCallback, useEffect, useRef, useState } from 'react'
+import { ReactNode, useEffect, useRef, useState } from 'react'
 
-import { getCmsClient } from '@cowprotocol/core'
 import { useWalletInfo } from '@cowprotocol/wallet'
 
 import styled from 'styled-components/macro'
 
+import { useTgAuthorization } from '../hooks/useTgAuthorization'
+import { useTgSubscription } from '../hooks/useTgSubscription'
 import { TelegramConnectionStatus } from '../pure/TelegramConnectionStatus'
-import { getTelegramAuth } from '../services/getTelegramAuth'
 
 const Wrapper = styled.div`
   padding-left: 12px;
@@ -21,124 +21,17 @@ const Option = styled.div`
   min-height: 40px;
 `
 
-interface TelegramData {
-  auth_date: number
-  first_name: string
-  hash: string
-  id: number
-  photo_url: string
-  username: string
-}
-
 const TELEGRAM_AUTH_WIDGET_URL = 'https://telegram.org/js/telegram-widget.js?22'
-const TG_BOT_ID = 7076584722 // cowNotificationsBot
 
-const AUTH_OPTIONS = {
-  bot_id: TG_BOT_ID,
-  lang: 'en',
-  request_access: 'write',
-}
-
-// TODO: Break down this large function into smaller functions
-// TODO: Add proper return type annotation
-// eslint-disable-next-line max-lines-per-function, @typescript-eslint/explicit-function-return-type
-export function ConnectTelegram() {
+export function ConnectTelegram(): ReactNode {
   const { account } = useWalletInfo()
-  const [isAuthChecked, setIsAuthChecked] = useState<boolean>(false)
   const [isTelegramScriptLoading, setIsTelegramScriptLoading] = useState<boolean>(true)
-  const [isTgSubscribed, setTgSubscribed] = useState<boolean>(false)
-  const [isCmsCallInProgress, setIsCmsCallInProgress] = useState<boolean>(false)
-  const [isLoginInProgress, setIsLoginInProgress] = useState<boolean>(false)
-
-  const [tgData, setTgData] = useState<TelegramData | null>(null)
   const telegramWrapperRef = useRef<HTMLDivElement>(null)
-  const isAuthRequestedRef = useRef(false)
 
-  // TODO: Add proper return type annotation
-  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-  const authenticate = (callback?: () => void) => {
-    getTelegramAuth(TG_BOT_ID, (response) => {
-      if (response && response.user) {
-        setTgData(response.user)
-      }
-      setIsAuthChecked(true)
-      callback?.()
-    })
-  }
+  const authorization = useTgAuthorization()
+  const { isTgSubscribed, isCmsCallInProgress, toggleSubscription } = useTgSubscription(account, authorization)
 
-  // TODO: Add proper return type annotation
-  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-  const authorize = (callback: () => void) => {
-    if (!window.Telegram) return
-
-    setIsLoginInProgress(true)
-
-    window.Telegram.Login.auth(AUTH_OPTIONS, (data) => {
-      if (data) {
-        setTgData(data)
-        setIsLoginInProgress(false)
-        callback()
-      } else {
-        authenticate(() => {
-          setIsLoginInProgress(false)
-          callback()
-        })
-      }
-    })
-  }
-
-  const checkOrAddTgSubscription = useCallback(
-    (method: string) => {
-      if (!tgData || !account) return
-
-      setIsCmsCallInProgress(true)
-
-      getCmsClient()
-        .POST(method, { body: { account, data: tgData } })
-        .then(({ data: result }: { data: boolean }) => {
-          setTgSubscribed(result)
-        })
-        .finally(() => {
-          setIsCmsCallInProgress(false)
-        })
-    },
-    [tgData, account],
-  )
-
-  // TODO: Add proper return type annotation
-  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-  const subscribeAccount = () => {
-    // TODO: Add proper return type annotation
-    // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-    const addSubscription = () => {
-      checkOrAddTgSubscription('/add-tg-subscription')
-    }
-
-    if (!isTgSubscribed) {
-      authorize(addSubscription)
-    } else {
-      addSubscription()
-    }
-  }
-
-  /**
-   * Check if the user is already authenticated
-   */
-  useEffect(() => {
-    if (isAuthRequestedRef.current) return
-
-    isAuthRequestedRef.current = true
-
-    authenticate()
-  }, [])
-
-  useEffect(() => {
-    if (!account || !tgData) return
-
-    setTgSubscribed(false)
-
-    checkOrAddTgSubscription('/check-tg-subscription')
-  }, [account, tgData, checkOrAddTgSubscription])
+  const { authorize, authenticate, tgData, isLoginInProgress, isAuthChecked } = authorization
 
   useEffect(() => {
     if (!telegramWrapperRef.current) return
@@ -158,7 +51,15 @@ export function ConnectTelegram() {
     telegramWrapperRef.current.appendChild(scriptElement)
   }, [])
 
+  /**
+   * Authenticate once on start
+   */
+  useEffect(() => {
+    authenticate()
+  }, [authenticate])
+
   const isLoading = isTelegramScriptLoading || !isAuthChecked || isCmsCallInProgress || isLoginInProgress
+  const needsAuthorization = isAuthChecked && !tgData
 
   return (
     <Wrapper ref={telegramWrapperRef}>
@@ -167,7 +68,9 @@ export function ConnectTelegram() {
         <TelegramConnectionStatus
           isLoading={isLoading}
           isSubscribed={isTgSubscribed}
-          subscribeAccount={subscribeAccount}
+          needsAuthorization={needsAuthorization}
+          authorize={authorize}
+          toggleSubscription={toggleSubscription}
         />
       </Option>
     </Wrapper>