+++ apps/cowswap-frontend/src/modules/trade/containers/TradeWidget/TradeWidgetModals.tsx
@@ -64,10 +64,12 @@ export function TradeWidgetModals({
   const updateApproveAmountState = useSetUserApproveAmountModalState()
 
   const resetAllScreens = useCallback(
-    (closeTokenSelectWidget = true) => {
+    (closeTokenSelectWidget = true, shouldCloseAutoImportModal = true) => {
       closeTradeConfirm()
       closeZeroApprovalModal()
-      closeAutoImportModal()
+      if (shouldCloseAutoImportModal) {
+        closeAutoImportModal()
+      }
       if (closeTokenSelectWidget) {
         updateSelectTokenWidgetState({ open: false })
       }
@@ -91,11 +93,11 @@ export function TradeWidgetModals({
   const error = tokenListAddingError || approveError || confirmError
 
   /**
-   * Close modals on chainId/account change
+   * Close all modals besides auto-import on account change
    */
   useEffect(() => {
-    resetAllScreens()
-  }, [chainId, account, resetAllScreens])
+    resetAllScreens(true, false)
+  }, [account, resetAllScreens])
 
   /**
    * Close all modals besides token select widget on chain change