+++ apps/explorer/src/components/orders/BridgeDetailsTable/BridgeDetailsContent/index.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode } from 'react'
 
 import { RECEIVED_LABEL } from '@cowprotocol/common-const'
-import { BridgeStatus, CrossChainOrder } from '@cowprotocol/cow-sdk'
+import { BridgeStatus, CrossChainOrder } from '@cowprotocol/sdk-bridging'
 
 import { AddressLink } from 'components/common/AddressLink'
 import { DetailRow } from 'components/common/DetailRow'
@@ -32,9 +32,10 @@ export function BridgeDetailsContent({ crossChainOrder }: BridgeDetailsContentPr
   const bridgeProvider = crossChainOrder.provider
   const { sourceToken, destinationToken } = useCrossChainTokens(crossChainOrder)
 
-  const RecipientAddress = (
+  const RecipientAddress = recipient ? (
     <AddressLink address={recipient} chainId={destinationChainId} bridgeProvider={bridgeProvider} showNetworkName />
-  )
+  ) : null
+
   return (
     <>
       <DetailRow label="Provider" tooltipText={BridgeDetailsTooltips.provider}>
@@ -52,7 +53,7 @@ export function BridgeDetailsContent({ crossChainOrder }: BridgeDetailsContentPr
       </DetailRow>
 
       <DetailRow label="To" tooltipText={BridgeDetailsTooltips.receiverAddress}>
-        <RowWithCopyButton textToCopy={recipient} contentsToDisplay={RecipientAddress} />
+        {recipient ? <RowWithCopyButton textToCopy={recipient} contentsToDisplay={RecipientAddress} /> : <span>-</span>}
       </DetailRow>
 
       <DetailRow label="Status" tooltipText={BridgeDetailsTooltips.status}>