+++ libs/common-const/src/bungeeAffiliateCode.ts
@@ -1,0 +1,2 @@
+export const bungeeAffiliateCode =
+  '609913096e1a3d62cecd0ffff47aa3e459eaedceb5fef75aad43e6cbff367039708902197e0b2b78b1d76cb0837ad0b318baedceb5fef75aad43e6cb'