+++ apps/cowswap-frontend/CHANGELOG.md
@@ -1,5 +1,82 @@
 # Changelog
 
+## [1.118.9](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.118.8...cowswap-v1.118.9) (2025-09-12)
+
+
+### Bug Fixes
+
+* ensure token selector closes on token selection ([#6252](https://github.com/cowprotocol/cowswap/issues/6252)) ([fe35182](https://github.com/cowprotocol/cowswap/commit/fe351822913fa8db44a34cc5c778ba4630da3e95))
+
+## [1.118.8](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.118.7...cowswap-v1.118.8) (2025-09-12)
+
+
+### Bug Fixes
+
+* skip trade if order valid time is non-positive ([#6243](https://github.com/cowprotocol/cowswap/issues/6243)) ([bc9783c](https://github.com/cowprotocol/cowswap/commit/bc9783c98be352047360b1e8b61b692aa309e464))
+* **swap:** improve lock screen logic in SwapWidget to prevent flash ([#6195](https://github.com/cowprotocol/cowswap/issues/6195)) ([159b199](https://github.com/cowprotocol/cowswap/commit/159b199e5e63d13feabe7832c0c8cd877024717a))
+
+## [1.118.7](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.118.6...cowswap-v1.118.7) (2025-09-11)
+
+
+### Bug Fixes
+
+* **twap:** fix navigation for big count of orders ([#6237](https://github.com/cowprotocol/cowswap/issues/6237)) ([dcf51a7](https://github.com/cowprotocol/cowswap/commit/dcf51a7dbc290ea9540d816e564d101daaed6dad))
+
+## [1.118.6](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.118.5...cowswap-v1.118.6) (2025-09-11)
+
+
+### Bug Fixes
+
+* **twap:** adjust safe tx history depth ([#6233](https://github.com/cowprotocol/cowswap/issues/6233)) ([3e626e1](https://github.com/cowprotocol/cowswap/commit/3e626e1b21df0102364f6670a092dbeda6e383d7))
+* **twap:** load orders history faster ([#6226](https://github.com/cowprotocol/cowswap/issues/6226)) ([bfba891](https://github.com/cowprotocol/cowswap/commit/bfba89170c63333e3182adb27624cac33106af1d))
+
+## [1.118.5](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.118.4...cowswap-v1.118.5) (2025-09-09)
+
+
+### Bug Fixes
+
+* **swap:** adjust swap banner header font size ([#6221](https://github.com/cowprotocol/cowswap/issues/6221)) ([c5d895d](https://github.com/cowprotocol/cowswap/commit/c5d895d2935737dbac9d4b0c9e71850f6011d0b5))
+
+## [1.118.4](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.118.3...cowswap-v1.118.4) (2025-09-09)
+
+
+### Bug Fixes
+
+* **swap:** replace swap unlock screen banner header ([#6217](https://github.com/cowprotocol/cowswap/issues/6217)) ([2e61b94](https://github.com/cowprotocol/cowswap/commit/2e61b942d5b31d14b8aa8884106a1206f6e85187))
+
+## [1.118.3](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.118.2...cowswap-v1.118.3) (2025-09-02)
+
+
+### Bug Fixes
+
+* **trade:** do not close auto-import modal when account changes ([#6209](https://github.com/cowprotocol/cowswap/issues/6209)) ([af643a9](https://github.com/cowprotocol/cowswap/commit/af643a94c0b28fe60639aeed22a69ea591b3fdcc))
+
+## [1.118.2](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.118.1...cowswap-v1.118.2) (2025-09-01)
+
+
+### Bug Fixes
+
+* do not crash when metadata is undefined ([#6204](https://github.com/cowprotocol/cowswap/issues/6204)) ([38de63a](https://github.com/cowprotocol/cowswap/commit/38de63ace9b104211e0eeacc4c68045fff6c793f))
+
+## [1.118.1](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.118.0...cowswap-v1.118.1) (2025-08-28)
+
+
+### Bug Fixes
+
+* **bungee:** remove custom affiliate logic/endpoints ([#6193](https://github.com/cowprotocol/cowswap/issues/6193)) ([d4feb5c](https://github.com/cowprotocol/cowswap/commit/d4feb5cbdedcbd4761498d0ab7c1e055cbc5e545))
+
+## [1.118.0](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.117.2...cowswap-v1.118.0) (2025-08-28)
+
+
+### Features
+
+* **account-proxy:** add links to cowshed releases ([#6175](https://github.com/cowprotocol/cowswap/issues/6175)) ([9641d53](https://github.com/cowprotocol/cowswap/commit/9641d5371ba5c7b682096386c2738991002506f5))
+* add swap unlock screen ([#6181](https://github.com/cowprotocol/cowswap/issues/6181)) ([6bb5593](https://github.com/cowprotocol/cowswap/commit/6bb55937792f9a13c620555a0a82eaeaa9d41e21))
+* **bridge:** add bridge appzi nps survey ([#6179](https://github.com/cowprotocol/cowswap/issues/6179)) ([83b8e2e](https://github.com/cowprotocol/cowswap/commit/83b8e2e1dec88af2e7f4b87162e3bce369244866))
+* **bridging:** enable mainnet-&gt;gnosis bridge ([#6172](https://github.com/cowprotocol/cowswap/issues/6172)) ([d1bb3d5](https://github.com/cowprotocol/cowswap/commit/d1bb3d5376cdd09ce2801a9bbf4a67b8b70f5d1b))
+* handle weekend for ondo tokens ([#6183](https://github.com/cowprotocol/cowswap/issues/6183)) ([cb3af07](https://github.com/cowprotocol/cowswap/commit/cb3af077192a24d09e23272fbcbfef69ea0f5812))
+* **trade:** implement partial approve ([#6139](https://github.com/cowprotocol/cowswap/issues/6139)) ([6f9d6b0](https://github.com/cowprotocol/cowswap/commit/6f9d6b055323d25705ee88a533f71190944be6e2))
+
 ## [1.117.2](https://github.com/cowprotocol/cowswap/compare/cowswap-v1.117.1...cowswap-v1.117.2) (2025-08-21)