+++ apps/explorer/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## [2.54.0](https://github.com/cowprotocol/cowswap/compare/explorer-v2.53.1...explorer-v2.54.0) (2025-09-12)
+
+
+### Features
+
+* restore Gnosis Affiliate header ([#6228](https://github.com/cowprotocol/cowswap/issues/6228)) ([22d0894](https://github.com/cowprotocol/cowswap/commit/22d08941a67032712e6d206ea8579786ed78715f))
+
+## [2.53.1](https://github.com/cowprotocol/cowswap/compare/explorer-v2.53.0...explorer-v2.53.1) (2025-08-28)
+
+
+### Bug Fixes
+
+* **bungee:** remove custom affiliate logic/endpoints ([#6193](https://github.com/cowprotocol/cowswap/issues/6193)) ([d4feb5c](https://github.com/cowprotocol/cowswap/commit/d4feb5cbdedcbd4761498d0ab7c1e055cbc5e545))
+
+## [2.53.0](https://github.com/cowprotocol/cowswap/compare/explorer-v2.52.0...explorer-v2.53.0) (2025-08-28)
+
+
+### Features
+
+* **bridging:** enable mainnet-&gt;gnosis bridge ([#6172](https://github.com/cowprotocol/cowswap/issues/6172)) ([d1bb3d5](https://github.com/cowprotocol/cowswap/commit/d1bb3d5376cdd09ce2801a9bbf4a67b8b70f5d1b))
+
+
+### Bug Fixes
+
+* **lens:** follow up fixes ([#6176](https://github.com/cowprotocol/cowswap/issues/6176)) ([cd8f3c0](https://github.com/cowprotocol/cowswap/commit/cd8f3c0201b8f04fc31a59588b6d2d05d215e112))
+
 ## [2.52.0](https://github.com/cowprotocol/cowswap/compare/explorer-v2.51.0...explorer-v2.52.0) (2025-08-21)