+++ libs/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.8.1](https://github.com/cowprotocol/cowswap/compare/core-v1.8.0...core-v1.8.1) (2025-09-11)
+
+
+### Bug Fixes
+
+* **twap:** load orders history faster ([#6226](https://github.com/cowprotocol/cowswap/issues/6226)) ([bfba891](https://github.com/cowprotocol/cowswap/commit/bfba89170c63333e3182adb27624cac33106af1d))
+
 ## [1.8.0](https://github.com/cowprotocol/cowswap/compare/core-v1.7.4...core-v1.8.0) (2025-08-21)